move all stark curve methods to main curve file
