+++ curve/curve.go
@@ -6,12 +6,18 @@ package curve
 	It is recommended to use in the same way(i.e. `curve.Sign` and not `ecdsa.Sign`).
 */
 import (
+	"bytes"
 	"crypto/elliptic"
+	"crypto/rand"
+	"crypto/sha256"
 	_ "embed"
 	"encoding/json"
 	"fmt"
 	"log"
 	"math/big"
+
+	junoCrypto "github.com/NethermindEth/juno/core/crypto"
+	"github.com/NethermindEth/juno/core/felt"
 )
 
 var Curve StarkCurve
@@ -271,16 +277,309 @@ func (sc StarkCurve) EcMult(m, x1, y1 *big.Int) (x, y *big.Int) {
 	return x, y
 }
 
-// Finds a nonnegative integer 0 <= x < p such that (m * x) % p == n
-//
-// (ref: https://github.com/starkware-libs/cairo-lang/blob/master/src/starkware/crypto/starkware/crypto/signature/math_utils.py)
-func DivMod(n, m, p *big.Int) *big.Int {
-	q := new(big.Int)
-	gx := new(big.Int)
-	gy := new(big.Int)
-	q.GCD(gx, gy, m, p)
+/*
+Verifies the validity of the stark curve signature
+given the message hash, and public key (x, y) coordinates
+used to sign the message.
 
-	r := new(big.Int).Mul(n, gx)
-	r = r.Mod(r, p)
-	return r
+(ref: https://github.com/starkware-libs/cairo-lang/blob/master/src/starkware/crypto/starkware/crypto/signature/signature.py)
+*/
+func (sc StarkCurve) Verify(msgHash, r, s, pubX, pubY *big.Int) bool {
+	w := sc.InvModCurveSize(s)
+
+	if s.Cmp(big.NewInt(0)) != 1 || s.Cmp(sc.N) != -1 {
+		return false
+	}
+	if r.Cmp(big.NewInt(0)) != 1 || r.Cmp(sc.Max) != -1 {
+		return false
+	}
+	if w.Cmp(big.NewInt(0)) != 1 || w.Cmp(sc.Max) != -1 {
+		return false
+	}
+	if msgHash.Cmp(big.NewInt(0)) != 1 || msgHash.Cmp(sc.Max) != -1 {
+		return false
+	}
+	if !sc.IsOnCurve(pubX, pubY) {
+		return false
+	}
+
+	zGx, zGy, err := sc.MimicEcMultAir(msgHash, sc.EcGenX, sc.EcGenY, sc.MinusShiftPointX, sc.MinusShiftPointY)
+	if err != nil {
+		return false
+	}
+
+	rQx, rQy, err := sc.MimicEcMultAir(r, pubX, pubY, sc.Gx, sc.Gy)
+	if err != nil {
+		return false
+	}
+	inX, inY := sc.Add(zGx, zGy, rQx, rQy)
+	wBx, wBy, err := sc.MimicEcMultAir(w, inX, inY, sc.Gx, sc.Gy)
+	if err != nil {
+		return false
+	}
+
+	outX, _ := sc.Add(wBx, wBy, sc.MinusShiftPointX, sc.MinusShiftPointY)
+	if r.Cmp(outX) == 0 {
+		return true
+	} else {
+		altY := new(big.Int).Neg(pubY)
+
+		zGx, zGy, err = sc.MimicEcMultAir(msgHash, sc.EcGenX, sc.EcGenY, sc.MinusShiftPointX, sc.MinusShiftPointY)
+		if err != nil {
+			return false
+		}
+
+		rQx, rQy, err = sc.MimicEcMultAir(r, pubX, new(big.Int).Set(altY), sc.Gx, sc.Gy)
+		if err != nil {
+			return false
+		}
+		inX, inY = sc.Add(zGx, zGy, rQx, rQy)
+		wBx, wBy, err = sc.MimicEcMultAir(w, inX, inY, sc.Gx, sc.Gy)
+		if err != nil {
+			return false
+		}
+
+		outX, _ = sc.Add(wBx, wBy, sc.MinusShiftPointX, sc.MinusShiftPointY)
+		if r.Cmp(outX) == 0 {
+			return true
+		}
+	}
+	return false
+}
+
+/*
+Signs the hash value of contents with the provided private key.
+Secret is generated using a golang implementation of RFC 6979.
+Implementation does not yet include "extra entropy" or "retry gen".
+
+(ref: https://datatracker.ietf.org/doc/html/rfc6979)
+*/
+func (sc StarkCurve) Sign(msgHash, privKey *big.Int, seed ...*big.Int) (x, y *big.Int, err error) {
+	if msgHash == nil {
+		return x, y, fmt.Errorf("nil msgHash")
+	}
+	if privKey == nil {
+		return x, y, fmt.Errorf("nil privKey")
+	}
+	if msgHash.Cmp(big.NewInt(0)) != 1 || msgHash.Cmp(sc.Max) != -1 {
+		return x, y, fmt.Errorf("invalid bit length")
+	}
+
+	inSeed := big.NewInt(0)
+	if len(seed) == 1 && inSeed != nil {
+		inSeed = seed[0]
+	}
+	for {
+		k := sc.GenerateSecret(big.NewInt(0).Set(msgHash), big.NewInt(0).Set(privKey), big.NewInt(0).Set(inSeed))
+		// In case r is rejected k shall be generated with new seed
+		inSeed = inSeed.Add(inSeed, big.NewInt(1))
+
+		r, _ := sc.EcMult(k, sc.EcGenX, sc.EcGenY)
+
+		// DIFF: in classic ECDSA, we take int(x) % n.
+		if r.Cmp(big.NewInt(0)) != 1 || r.Cmp(sc.Max) != -1 {
+			// Bad value. This fails with negligible probability.
+			continue
+		}
+
+		agg := new(big.Int).Mul(r, privKey)
+		agg = agg.Add(agg, msgHash)
+
+		if new(big.Int).Mod(agg, sc.N).Cmp(big.NewInt(0)) == 0 {
+			// Bad value. This fails with negligible probability.
+			continue
+		}
+
+		w := DivMod(k, agg, sc.N)
+		if w.Cmp(big.NewInt(0)) != 1 || w.Cmp(sc.Max) != -1 {
+			// Bad value. This fails with negligible probability.
+			continue
+		}
+
+		s := sc.InvModCurveSize(w)
+		return r, s, nil
+	}
+
+	return x, y, nil
+}
+
+/*
+See Sign. SignFelt just wraps Sign.
+*/
+func (sc StarkCurve) SignFelt(msgHash, privKey *felt.Felt) (*felt.Felt, *felt.Felt, error) {
+	msgHashInt := msgHash.BigInt(new(big.Int))
+	privKeyInt := privKey.BigInt(new(big.Int))
+	x, y, err := sc.Sign(msgHashInt, privKeyInt)
+	if err != nil {
+		return nil, nil, err
+	}
+	xFelt := felt.NewFelt(new(felt.Felt).Impl().SetBigInt(x))
+	yFelt := felt.NewFelt(new(felt.Felt).Impl().SetBigInt(y))
+	return xFelt, yFelt, nil
+
+}
+
+/*
+Hashes the contents of a given array using a golang Pedersen Hash implementation.
+
+(ref: https://github.com/seanjameshan/starknet.js/blob/main/src/utils/ellipticCurve.ts)
+*/
+func (sc StarkCurve) HashElements(elems []*big.Int) (hash *big.Int, err error) {
+	if len(elems) == 0 {
+		elems = append(elems, big.NewInt(0))
+	}
+
+	hash = big.NewInt(0)
+	for _, h := range elems {
+		hash, err = sc.PedersenHash([]*big.Int{hash, h})
+		if err != nil {
+			return hash, err
+		}
+	}
+	return hash, err
+}
+
+/*
+Hashes the contents of a given array with its size using a golang Pedersen Hash implementation.
+
+(ref: https://github.com/starkware-libs/cairo-lang/blob/13cef109cd811474de114925ee61fd5ac84a25eb/src/starkware/cairo/common/hash_state.py#L6)
+*/
+func (sc StarkCurve) ComputeHashOnElements(elems []*big.Int) (hash *big.Int, err error) {
+	elems = append(elems, big.NewInt(int64(len(elems))))
+	return Curve.HashElements((elems))
+}
+
+/*
+Provides the pedersen hash of given array of big integers.
+NOTE: This function assumes the curve has been initialized with contant points
+
+(ref: https://github.com/seanjameshan/starknet.js/blob/main/src/utils/ellipticCurve.ts)
+*/
+func (sc StarkCurve) PedersenHash(elems []*big.Int) (hash *big.Int, err error) {
+	if len(sc.ConstantPoints) == 0 {
+		return hash, fmt.Errorf("must initiate precomputed constant points")
+	}
+
+	ptx := new(big.Int).Set(sc.Gx)
+	pty := new(big.Int).Set(sc.Gy)
+	for i, elem := range elems {
+		x := new(big.Int).Set(elem)
+
+		if x.Cmp(big.NewInt(0)) != -1 && x.Cmp(sc.P) != -1 {
+			return ptx, fmt.Errorf("invalid x: %v", x)
+		}
+
+		for j := 0; j < 252; j++ {
+			idx := 2 + (i * 252) + j
+			xin := new(big.Int).Set(sc.ConstantPoints[idx][0])
+			yin := new(big.Int).Set(sc.ConstantPoints[idx][1])
+			if xin.Cmp(ptx) == 0 {
+				return hash, fmt.Errorf("constant point duplication: %v %v", ptx, xin)
+			}
+			if x.Bit(0) == 1 {
+				ptx, pty = sc.Add(ptx, pty, xin, yin)
+			}
+			x = x.Rsh(x, 1)
+		}
+	}
+
+	return ptx, nil
+}
+
+/*
+Provides the pedersen hash of given array of felts.
+NOTE: This function just wraps the Juno implementation
+
+(ref: https://github.com/NethermindEth/juno/blob/main/core/crypto/poseidon_hash.go#L74)
+*/
+func (sc StarkCurve) PoseidonArray(felts ...*felt.Felt) *felt.Felt {
+	return junoCrypto.PoseidonArray(felts...)
+}
+
+/*
+Provides the starknet keccak hash .
+NOTE: This function just wraps the Juno implementation
+
+(ref: https://github.com/NethermindEth/juno/blob/main/core/crypto/keccak.go#L11)
+*/
+func (sc StarkCurve) StarknetKeccak(b []byte) (*felt.Felt, error) {
+	return junoCrypto.StarknetKeccak(b)
+}
+
+// implementation based on https://github.com/codahale/rfc6979/blob/master/rfc6979.go
+// for the specification, see https://tools.ietf.org/html/rfc6979#section-3.2
+func (sc StarkCurve) GenerateSecret(msgHash, privKey, seed *big.Int) (secret *big.Int) {
+	alg := sha256.New
+	holen := alg().Size()
+	rolen := (sc.BitSize + 7) >> 3
+
+	if msgHash.BitLen()%8 <= 4 && msgHash.BitLen() >= 248 {
+		msgHash = msgHash.Mul(msgHash, big.NewInt(16))
+	}
+
+	by := append(int2octets(privKey, rolen), bits2octets(msgHash, sc.N, sc.BitSize, rolen)...)
+
+	if seed.Cmp(big.NewInt(0)) == 1 {
+		by = append(by, seed.Bytes()...)
+	}
+
+	v := bytes.Repeat([]byte{0x01}, holen)
+
+	k := bytes.Repeat([]byte{0x00}, holen)
+
+	k = mac(alg, k, append(append(v, 0x00), by...), k)
+
+	v = mac(alg, k, v, v)
+
+	k = mac(alg, k, append(append(v, 0x01), by...), k)
+
+	v = mac(alg, k, v, v)
+
+	for {
+		var t []byte
+
+		for len(t) < rolen {
+			v = mac(alg, k, v, v)
+			t = append(t, v...)
+		}
+
+		secret = bits2int(new(big.Int).SetBytes(t), sc.BitSize)
+		// TODO: implement seed here, final gating function
+		if secret.Cmp(big.NewInt(0)) == 1 && secret.Cmp(sc.N) == -1 {
+			return secret
+		}
+		k = mac(alg, k, append(v, 0x00), k)
+		v = mac(alg, k, v, v)
+	}
+}
+
+// obtain random primary key on stark curve
+// NOTE: to be used for testing purposes
+func (sc StarkCurve) GetRandomPrivateKey() (priv *big.Int, err error) {
+	max := new(big.Int).Sub(sc.Max, big.NewInt(1))
+
+	priv, err = rand.Int(rand.Reader, max)
+	if err != nil {
+		return priv, err
+	}
+
+	x, y, err := sc.PrivateToPoint(priv)
+	if err != nil {
+		return priv, err
+	}
+
+	if !sc.IsOnCurve(x, y) {
+		return priv, fmt.Errorf("key gen is not on stark cruve")
+	}
+
+	return priv, nil
+}
+
+// obtain public key coordinates from stark curve given the private key
+func (sc StarkCurve) PrivateToPoint(privKey *big.Int) (x, y *big.Int, err error) {
+	if privKey.Cmp(big.NewInt(0)) != 1 || privKey.Cmp(sc.N) != -1 {
+		return x, y, fmt.Errorf("private key not in curve range")
+	}
+	x, y = sc.EcMult(privKey, sc.EcGenX, sc.EcGenY)
+	return x, y, nil
 }
